Option including delimiters in precidence regexes
The option defaults to false, but the use-case is for a coding standard
with a preferred include order like:

    - stdafx.h (for precompiled header – must be first).
    - myfilename.h (corresponding header for .cpp).
    - C system files.
    - C++ system files.
    - Other libraries' .h files.
    - Your project's .h files.

In the above case, C system files can be detected as includes using angled
brackets and containing a dot (e.g. '<[^.]*\.'), while C++ system files
can be detected as includes using angled brackets but contining no dot
(e.g.  '<'). So, to make this work, the delimiters must be included for
comparison in the regexes.